[760 тыс. ключей в секунду]: оптимизация редукции Монтгомери за счёт сокращённого умножения

+++ test.h
@@ -15,4 +15,4 @@ struct Timer
 
 bool check(unsigned char* array1, unsigned char* array2, int length);
 int round(double number, int digits);
-long long test();
+int test();
